LayoutTests/media/video-element-fullscreen-not-in-dom-accelerated-iphone-expected.txt

Test that playing a video in fullscreen is accelerated.

EVENT(webkitbeginfullscreen)
EXPECTED (window.internals.mediaPlayerRenderingCanBeAccelerated(video) == 'true') OK
END OF TEST
